Fix decoding internal server errors
When receiving a 500 Internal Server Error response the problem decoder
would try to decode that into a problems.InternalProblem. This problem
struct will try to use the wrapped error cause when producing a
formatted error message. Since this field is excluded when marshaling
the struct it will be nil when decoding to it, cause a nil pointer panic
to be printed in the error message.

Decode these issues to a problems.BasicProblem instead.